refactor: don't include test ids for prod
