+++ ui/src/pages/clients/ClientCreate/index.ts
@@ -1,2 +1,2 @@
 export { default } from "./ClientCreate";
-export { TestId as ClientCreateTestId } from "./types";
+export { TestId as ClientCreateTestId } from "./test-types";
