+++ ui/src/pages/clients/ClientEdit/index.ts
@@ -1,2 +1,2 @@
 export { default } from "./ClientEdit";
-export { TestId as ClientEditTestId } from "./types";
+export { TestId as ClientEditTestId } from "./test-types";
