+++ ui/src/pages/identities/IdentityCreate/index.ts
@@ -1,2 +1,2 @@
 export { default } from "./IdentityCreate";
-export { TestId as IdentityCreateTestId } from "./types";
+export { TestId as IdentityCreateTestId } from "./test-types";
