+++ ui/src/pages/providers/ProviderEdit/index.ts
@@ -1,2 +1,2 @@
 export { default } from "./ProviderEdit";
-export { TestId as ProviderEditTestId } from "./types";
+export { TestId as ProviderEditTestId } from "./test-types";
